requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django_extensions==1.9.9
gevent==1.2.2
gunicorn==19.7.1
psycopg2==2.7.3.2
